Refactored initialization. Removed bashpodder compatibility
Bashpodder support is no longer that useful to allow people to migrate, and
migration is still possible by moving a few files around.

Initialization has been refactored into a single subroutine which can
be called any number of times without it re-initializing over and over.
It also greatly simplifies a bunch of the initialization logic.