Added a metadata store and replaced the podcast log
The log is now in downloaded.json, which contains additional metadata such as
which feed the file was downloaded from and when it was downloaded.

The metadata storage is in meta.json, which contains information about which
files belong to which feeds, as well as the titles we have known for each of
the feeds, which should make catalogue entries more consistent in time.